Door Height + Capacity
Fixed a pathfinding issue caused by doors being really high, and made building capacity actually affect things.

+++ Assets/Scripts/Building.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Building : MonoBehaviour {
 
 	public Transform inside;
-	public int capacity;
+	public float capacity;
     public List<GameObject> doors;
 
     [Header("Change Per Second")]
